A revised ASPNetImage library with these news properties and methods: AddImage AntiAliasText BackgroundColor BrightenImage Bold Contrast CreateGrayScale CreateNegative CropImage DarkenImage Emboss FileName FontColor FontName FontSize ImageFormat JPEGQuality LoadBlob LoadImage MaxX MaxY Resize ResizeR RotateImage SaveImage TextOut TextHeight TextWidth
It fully replaces the original AspImage.dll file, so no modifications to the ASP source code is required. It is obtained simply adding the [PrgID] annotation in the main .cs file.

In the Resource folder you can find SearchProgID.vbs file which is useful to know what dll is actually installed on a given server.